notification_service/requirements.txt
fastapi
uvicorn
asyncpg
aiosmtplib
telethon
httpx
python-dotenv
